backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==23.1.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
Django==4.2.16
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==4.5.0
django-environ==0.11.2
django-rest-swagger==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.1
msgpack==1.1.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pip-chill==1.0.3
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.2.1
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==75.2.0
simplejson==3.19.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
zope.interface==7.1.0

